+++ examples/rust/exercises/part-3-sol/src/lib.rs
@@ -22,7 +22,7 @@ mod logic;
 mod serial;
 
 fn main() {
-    let uart = uart::Uart(0);
+    let uart = uart::Uart::new(0).unwrap();
     let mut serial = Serial::new(&uart);
     let mut logic = Logic::new();
     loop {