+++ examples/rust/exercises/part-4/src/lib.rs
@@ -25,7 +25,7 @@ mod serial;
 mod store; // TODO: Implement this module.
 
 fn main() {
-    let uart = uart::Uart(0);
+    let uart = uart::Uart::new(0).unwrap();
     let mut serial = Serial::new(&uart);
     let mut logic = Logic::new();
     loop {