+++ examples/rust/exercises/part-7/src/lib.rs
@@ -29,7 +29,7 @@ mod store;
 mod touch; // TODO: Update this module.
 
 fn main() {
-    let uart = uart::Uart(0);
+    let uart = uart::Uart::new(0).unwrap();
     let mut serial = Serial::new(&uart);
     let mut logic = Logic::new();
     loop {